x/likenft/client/cli/query_claimable_nft.go
package cli

import (
	"context"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/likecoin/likechain/x/likenft/types"
	"github.com/spf13/cobra"
)

func CmdListMintableNFT() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "mintable-nft-index",
		Short: "Enumerate all Mintable NFT Contents under all classes",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryMintableNFTIndexRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.MintableNFTIndex(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowMintableNFT() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "mintable-nft [class-id] [mintable-id]",
		Short: "Query a specific Mintable NFT Content",
		Args:  cobra.ExactArgs(2),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			argClassId := args[0]
			argId := args[1]

			params := &types.QueryMintableNFTRequest{
				ClassId: argClassId,
				Id:      argId,
			}

			res, err := queryClient.MintableNFT(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdMintableNFTs() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "mintable-nfts [class-id]",
		Short: "Query Mintable NFT Contents under a class",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			reqClassId := args[0]

			clientCtx, err := client.GetClientTxContext(cmd)
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryMintableNFTsRequest{

				ClassId: reqClassId,
			}

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}
			params.Pagination = pageReq

			res, err := queryClient.MintableNFTs(cmd.Context(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
